+++ app/client/cli/cmd.go
@@ -2,23 +2,34 @@ package cli
 
 import (
 	"context"
+	"log"
+	"os"
 
+	// NOTE: Imported for debug purposes in order to populate the keybase with the pre-generated keys
+	_ "github.com/pokt-network/pocket/app/client/keybase/debug"
 	"github.com/pokt-network/pocket/runtime/defaults"
 	"github.com/spf13/cobra"
 )
 
-const cliExecutableName = "client"
+const (
+	cliExecutableName = "client"
+	keybaseSuffix     = "/keys"
+)
 
 var (
-	remoteCLIURL       string
-	privateKeyFilePath string
-	nonInteractive     bool
+	remoteCLIURL   string
+	dataDir        string
+	nonInteractive bool
 )
 
 func init() {
+	homeDir, err := os.UserHomeDir()
+	if err != nil {
+		log.Fatalf("[ERROR] Cannot find user home directory: %s", err.Error())
+	}
 	rootCmd.PersistentFlags().StringVar(&remoteCLIURL, "remote_cli_url", defaults.DefaultRemoteCLIURL, "takes a remote endpoint in the form of <protocol>://<host> (uses RPC Port)")
-	rootCmd.PersistentFlags().StringVar(&privateKeyFilePath, "path_to_private_key_file", "./pk.json", "Path to private key to use when signing")
 	rootCmd.PersistentFlags().BoolVar(&nonInteractive, "non_interactive", false, "if true skips the interactive prompts wherever possible (useful for scripting & automation)")
+	rootCmd.PersistentFlags().StringVar(&dataDir, "data_dir", homeDir+"/.pocket", "Path to store pocket related data (keybase etc.)")
 }
 
 var rootCmd = &cobra.Command{