+++ app/client/doc/CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.12] - 2023-02-10
+## [0.0.0.13] - 2023-02-15
 
 - Added `not_interactive` flag to allow for non-interactive `Stake` and `Unstake` transactions (dogfooding in `cluster-manager`)
 - Updated CLI to use to source the address book and the current height from the RPC server leveraging the `rpcAddressBookProvider` and `rpcCurrentHeightProvider` respectively and the `bus` for dependency injection
+
+## [0.0.0.12] - 2023-02-14
+
+- Integrate keybase with CLI
+- Add debug module to keybase to automatically populate keybase with 999 validators
 
 ## [0.0.0.11] - 2023-02-09